requirements.txt
# Core dependencies
pyyaml>=6.0.1
requests>=2.31.0
python-dotenv>=1.0.0

# LangChain and AI integrations
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-core>=0.1.0
